Clean up plot_gallery modules.
It appears to be common practice in Python to put related classes in the same module file. This certainly makes the module pathing less awkward in my plot_gallery. However, as someone new to the language who prefers small source code files, this seems to be an odd pattern.

Also, the reason the diffs are so large when regenerating the Jupyter notebook is I think due to the <dc:date> timestamps changing in the SVG's' produced by Matplotlib. I will have to see if there is an easy way to normalize these out.